Updated controllers to expose new endpoints for Delete methods using primary keys

+++ DocumentDataAPI/DocumentDataAPI/Data/Repositories/IWordRatioRepository.cs
@@ -4,8 +4,8 @@ namespace DocumentDataAPI.Data.Repositories;
 
 public interface IWordRatioRepository : IRepository<WordRatioModel>
 {
-    Task<WordRatioModel?> Get(int id, string word);
-    Task<int> Delete(int id, string word);
+    Task<WordRatioModel?> Get(long id, string word);
+    Task<int> Delete(long id, string word);
     Task<IEnumerable<WordRatioModel>> GetByWord(string word);
     Task<IEnumerable<WordRatioModel>> GetByWords(IEnumerable<string> word);
     Task<IEnumerable<WordRatioModel>> GetByDocumentId(int id);
